Task done!
In the first part of this task (index1, style1, script1), authorization by e-mail, password and password confirmation is implemented. After successful completion and verification, the user's mail is stored in memory and transferred to the next page with a biography.

In the second part of the task (index2, style2, script2), the user's biography is filled in: first name, last name, year of birth, gender, (optional) phone number, (optional) skype. After pressing the "Save" button, the correctness of filling in the data is checked and it is written to memory. When you click on the "exit" button, the current session ends, the e-mail is deleted from memory, and all data remains.

Data is stored in local storage.